app/requirements.txt
Django==2.2.13
gunicorn==20.0.4
psycopg2-binary==2.8.5
django-grappelli==2.13.4
django-filebrowser==3.12.1
sorl-thumbnail==12.6.3
django-simple-captcha==0.5.12
django-autotranslate==1.1.1
django-chunked-upload==2.0.0
django-localflavor==3.0.1
django-tinymce==3.0.1
django-crispy-forms==1.9.2
django-countries==6.1.2
django-phonenumber-field==4.0.0
django-extensions==3.0.3
python-dateutil==2.8.1
phonenumbers==8.12.6
GDAL==2.4
pylzma==0.5
pykml==0.2
numpy==1.19.1
redis==3.5.3
celery[redis]==4.4.7
simplekml==1.3.5